Add functions to deal with HTTP dates
HTTP 1.1 uses the date format described in RFC 1123, and Twitter
uses the same date format to timestamp the statuses.

How I would like for HTTP to use the ISO-8601 date format instead.

At the moment, we are rolling out our own version. We should
use SoupDate instead, but as a first approximation this isn't too
bad.