Implement get_friends(), get_followers() and show_user()
Implement the declared but missing methods of TwitterClient.

The get_friends() and get_followers() methods have been updated with the
"omit-status" argument, that returns the TwitterUser without the status
field set. This variant should be used to avoid spending time retrieving
and parsing the status for the cases where just the user list is needed.