ClutterModel implementation holding TwitterStatus
A specialized implementation of a ClutterModel base class,
using GSequence to store a list of TwitterStatus. Might have
used a GPtrArray, as the list is sorted using the same
ordering given us by the web service, but I didn't want to
rewrite everything.

TweetStatusModel holds a single column, because all the data
is really stored inside the TwitterStatus object and there's
no point in duplicating it.