Add a simple animation API
TweetAnimation is an experimental animation API that supercedes the
effects API in Clutter. It replaces all the effects functions with:

  tweet_actor_animate (actor, animation_mode, duration,
                       "property-1", interval_1,
                       "property-2", interval_2,
                       "property-3", interval_3,
                       NULL);

Which will animate the three listed properties within the given
intervals.

The animation API does not depend on internals, and wraps timelines
and alpha functions internally, so it's reusing the Clutter animation
framework as much as possible.