Add the in-reply-to accessors to TwitterStatus
A Twitter status can be in reply of a user or a status from a user.
Twitter will will two fields in the status structure with the ID of
the user and the status we a replying to - so we can link or retrieve
both.

TwitterStatus now parses those fields has the API to access that
data.