Add cell geometry accessor function
Sometimes is useful to know the geometry of a cell inside a
ListView given the (row, column) pair. The newly added:

  tidy_list_view_get_cell_geometry()

function computes the allocated geometry of a cell, given
its row and column; it can also adjust the position depending
on the current value of the adjustments, in case the ListView
is inside a TidyScrollable.