Add initial status info overlay actor
Implement the StatusInfo actor, which takes a TwitterStatus and
shows more informations than what the StatusView does.

At the moment, it just shows a bigger profile image and the same
status text, but it will get:

 - whether the status is a reply to someone
 - a button to set the status as a favourite
 - a button to reply to the status

and possibly a way to open the status in the browser and to delete
a status (if it belongs to us).