Make the cell a little bit shorter
The default height of the StatusCell is 84px instead of 108px; this
allows more cells to be displayed at the same time, but we are getting
really close to the text label size at Sans 8.

I need to make the cell dinamically allocate enough space to compensate
a longer text, and use DEFAULT_HEIGHT as the minimum height.