Try to avoid double-tapping issues
If the user double-taps on the status view, the first
button-press event will reach the status view but the
second will probably reach the StatusInfo actor, closing
it as soon as it's open.

To avoid this issue, we set the StatusInfo as non-reactive
until it's fully visible (the animation has emitted the
::completed signal); we also keep a pointer to the current
StatusInfo overlay inside the TweetWindow private data
structure.