Build the preferences dialog
Wrap the preferences dialog constructed using GtkBuilder behind a
single function call.

When the preferences dialog has been created, attach it to the
parent window passed to tweet_show_preferences_dialog() and hide it
on ::response. When the parent window has been destroyed, it will
bring the preferences dialog with her.