Keep track of network state changes
Simply refreshing the StatusView every time the "active-connection"
state is received creates wrong renderings of the status list.

TweetWindow has to store the current network state and avoid queuing
up work if the new network state matches the old. In case the network
state changes to "active-connection" and we are still loading from
a previous state, do not attach the timeout handler twice but always
check for the source id being set first.