Keep updating instead of rebuilding the statuses
Instead of refreshing the whole statuses model, just get the statuses
that were added since the last update. This should keep the network
traffic down, the list view update fast and allow us to further on
notify the user about the new statues without keeping an internal
list of statuses.