Avoid adding statuses already in the model
Since statuses have a unique id, adding a status already inside the
model should always check for duplicates.

Right now the check is not at all optimal: it's a full lookup on the
GSequence for each insertion; I should probably keep an hash of all
the statuses using the id as the key, and use that for the lookups.