Get the user from Twitter before the user's timeline
We need to keep the user around - for checking user preferences, location
and other stuff we might want to update and/or display. Also having the
user id would make looking up the archive and other data a lot easier.

Since it's better to serialize the requests to Twitter to avoid the
failwhale to fall on our heads followed by a vase of petunias, we get the
user info first and then we ask for the user's timeline if the action was
successfull.