Work around the allocation weirdness in ClutterLabel
ClutterLabel in the 0.6 releases has weird allocation issues caused
by the fact that we don't have a way to get the preferred size of an
actor. This will be fixed by the new size negotiation API in Clutter
0.8, but for the time being is necessary to set the height of a
Label to 1 when setting the width, in order to retrieve the right
height later on.

This fixes the wrong size of the speech bubbles containing the
Twitter status.

Signed-off-by: Emmanuele Bassi <ebassi@openedhand.com>