Add TweetCanvas, a wrapper for GtkClutterEmbed
GtkClutterEmbed doesn't have a size-request implementation, mostly
because we don't want to force a minimum size. This might actually
change in Clutter 0.8 because we have a way to ask the scenegraph
for its minimum size.

To make a long story short: Tweet has a minimum size, and it can
use it; to avoid nasty crap like gtk_widget_set_size_request() and
to add a more indirection for when we switch over to the new
goodness currently in trunk, we provide a new GtkWidget wrapping
GtkClutterEmbed and providing:

  - a size-request implementation, with a minimum size
  - a border-width property, that we can use to pad the finger
    scroll view